respawneo de la rana en la posición leída por archivo

+++ src/game.h
@@ -72,7 +72,7 @@ public:
 	Collision checkCollision(const SDL_FRect& rect) const;
 
 	//Devuelve la posición de spawn de la rana
-	Point2D<int> getFrogSpawn() const;
+	Point2D<float> getFrogSpawn() const;
 };
 
 inline Texture*
